requirements.txt
python-telegram-bot==20.7
telethon==1.30.0
aiohttp==3.9.5